+++ Raymond/Constants.h
@@ -1,6 +1,6 @@
 #ifndef H_RAYMOND_CONSTANTS
 #define H_RAYMOND_CONSTANTS
 
-static const float EPSILON = 0.0001f;
+static const double EPSILON = 0.0001;
 
 #endif
