+++ Raymond/Material.cpp
@@ -40,8 +40,8 @@ Color NormalsMaterial::lighting(std::shared_ptr<Light>, IxComps & comps) const
 {
 	// returns the normal as a color, mapping from (-1.0, 1.0) to (0.0, 1.0)
 	// Assumes a normalized vector
-	Tuple n = comps.normal_v - Tuple::Vector(0.0f, 0.0f, -1.0f);
-	return Color(n.x + 1.0f, n.y + 1.0f, n.z + 1.0f) * 0.5f;
+	Tuple n = comps.normal_v - Tuple::Vector(0.0, 0.0, -1.0);
+	return Color(n.x + 1.0, n.y + 1.0, n.z + 1.0) * 0.5;
 }
 
 
@@ -57,11 +57,11 @@ PhongMaterial::PhongMaterial() : BaseMaterial()
 {
 	this->name = "Default Phong Material 000";
 
-	this->color = Color(1.0f, 1.0f, 1.0f);
+	this->color = Color(1.0, 1.0, 1.0);
 
-	this->ambient = 0.1f;
-	this->diffuse = 0.9f;
-	this->specular = 0.9f;
+	this->ambient = 0.1;
+	this->diffuse = 0.9;
+	this->specular = 0.9;
 	this->shininess = 200.0;
 }
 
@@ -77,9 +77,9 @@ Color PhongMaterial::lighting(std::shared_ptr<Light> lgt, IxComps & comps) const
 {
 	Color effective_col, ambient_col, diffuse_col, specular_col;
 	Tuple light_v, reflect_v;
-	float light_dot_normal, reflect_dot_eye, factor;
+	double light_dot_normal, reflect_dot_eye, factor;
 
-	Color black = Color(0.0f, 0.0f, 0.0f);
+	Color black = Color(0.0, 0.0, 0.0);
 
 	// Combine the surface color with the light's color and intensity
 	effective_col = this->color * lgt->color;
@@ -94,7 +94,7 @@ Color PhongMaterial::lighting(std::shared_ptr<Light> lgt, IxComps & comps) const
 	// the light is on the other side of the surface.
 	light_dot_normal = Tuple::dot(light_v, comps.normal_v);
 
-	if (light_dot_normal < 0.0f || comps.shadow_multiplier < lgt->cutoff)
+	if (light_dot_normal < 0.0 || comps.shadow_multiplier < lgt->cutoff)
 	{
 		diffuse_col = black;
 		specular_col = black;
@@ -106,7 +106,7 @@ Color PhongMaterial::lighting(std::shared_ptr<Light> lgt, IxComps & comps) const
 
 		reflect_dot_eye = Tuple::dot(reflect_v, comps.eye_v);
 
-		if (reflect_dot_eye <= 0.0f)
+		if (reflect_dot_eye <= 0.0)
 		{
 			specular_col = black;
 		}
@@ -149,7 +149,7 @@ Color PhongMaterial::lighting(
 	comps.point = point;
 	comps.eye_v = eye_v;
 	comps.normal_v = normal_v;
-	comps.shadow_multiplier = shadowed ? 0.0f : 1.0f;
+	comps.shadow_multiplier = shadowed ? 0.0 : 1.0;
 
 	return this->lighting(lgt, comps);
 }
