+++ Raymond/Matrix.h
@@ -12,8 +12,8 @@ class Matrix
 {
 public:
 	// Constructors
-	Matrix(int side, float fill);
-	Matrix(int side, std::vector<float> values);
+	Matrix(int side, double fill);
+	Matrix(int side, std::vector<double> values);
 	Matrix(int);
 	~Matrix();
 
@@ -22,45 +22,45 @@ public:
 
 	//Methods
 	// Accessors
-	float get(int row, int col) const;
-	float& get(int index);
-	float get(int index) const;
-	float& at(int index);
-	float at(int index) const;
+	double get(int row, int col) const;
+	double& get(int index);
+	double get(int index) const;
+	double& at(int index);
+	double at(int index) const;
 
-	virtual std::vector<float> get_row(int row) const;
-	virtual std::vector<float> get_column(int col) const;
+	virtual std::vector<double> get_row(int row) const;
+	virtual std::vector<double> get_column(int col) const;
 
 	int get_num_columns() const;
 	int get_num_rows() const;
 
-	void set(int row, int col, float value);
-	void set(int index, float value);
-	void set_multiple(const std::vector<float> values);
+	void set(int row, int col, double value);
+	void set(int index, double value);
+	void set_multiple(const std::vector<double> values);
 
 	void generate_identity();
 
-	std::vector<float> transpose_vector() const;
+	std::vector<double> transpose_vector() const;
 	Matrix transpose() const;
-	std::vector<float> sub_matrix_vector(int remove_row, int remove_col) const;
+	std::vector<double> sub_matrix_vector(int remove_row, int remove_col) const;
 	Matrix sub_matrix(int remove_row, int remove_col) const;
-	float minor(int row, int col) const;
-	float cofactor(int row, int col) const;
-	float determinant() const;
+	double minor(int row, int col) const;
+	double cofactor(int row, int col) const;
+	double determinant() const;
 	bool is_invertable() const;
-	std::vector<float> inverse_vector() const;
+	std::vector<double> inverse_vector() const;
 	Matrix inverse() const;
 
 	std::string to_string() const;
 
 	// iterators
-	float * begin();
-	const float * begin() const;
-	float * end();
-	const float * end() const;
+	double * begin();
+	const double * begin() const;
+	double * end();
+	const double * end() const;
 
 	// Overloaded Operators
-	float& operator[](int index);
+	double& operator[](int index);
 
 	// Friendly Overloaded Operators
 	friend std::ostream & operator<<(std::ostream & ox, const Matrix & m);
@@ -69,7 +69,7 @@ public:
 
 private:
 	// Properties
-	std::vector<float> m_data_;
+	std::vector<double> m_data_;
 	int m_num_columns_, m_num_rows_;
 
 	//Methods
@@ -82,16 +82,16 @@ class Matrix2 :
 public:
 	// Constructors
 	Matrix2();
-	Matrix2(float fill);
-	Matrix2(const std::vector<float> v);
+	Matrix2(double fill);
+	Matrix2(const std::vector<double> v);
 
 	// Factories
 	static Matrix2 Identity();
 
 	// Methods
-	std::vector<float> get_row(int row) const;
-	std::vector<float> get_column(int col) const;
-	float determinant() const;
+	std::vector<double> get_row(int row) const;
+	std::vector<double> get_column(int col) const;
+	double determinant() const;
 	Matrix2 inverse() const;
 	Matrix2 transpose() const;
 };
@@ -102,20 +102,20 @@ class Matrix3 :
 public:
 	// Constructors
 	Matrix3();
-	Matrix3(float fill);
-	Matrix3(const std::vector<float> v);
+	Matrix3(double fill);
+	Matrix3(const std::vector<double> v);
 
 	// Factories
 	static Matrix3 Identity();
 
 	// Methods
-	std::vector<float> get_row(int row) const;
-	std::vector<float> get_column(int col) const;
+	std::vector<double> get_row(int row) const;
+	std::vector<double> get_column(int col) const;
 
 	Matrix2 sub_matrix3(int remove_row, int remove_col) const;
-	float minor(int row, int col) const;
-	float cofactor(int row, int col) const;
-	float determinant() const;
+	double minor(int row, int col) const;
+	double cofactor(int row, int col) const;
+	double determinant() const;
 	Matrix3 inverse() const;
 	Matrix3 transpose() const;
 };
@@ -126,34 +126,34 @@ class Matrix4 :
 public:
 	// Constructors
 	Matrix4();
-	Matrix4(float fill);
-	Matrix4(const std::vector<float> v);
+	Matrix4(double fill);
+	Matrix4(const std::vector<double> v);
 
 	// Factories
 	static Matrix4 Identity();
 
 	// Transformation Matrix Factories
-	static Matrix4 Translation(float x, float y, float z);
+	static Matrix4 Translation(double x, double y, double z);
 	static Matrix4 Translation(Tuple p);
-	static Matrix4 Scaling(float x, float y, float z);
+	static Matrix4 Scaling(double x, double y, double z);
 	static Matrix4 Scaling(Tuple s);
-	static Matrix4 Rotation_X(float radians);
-	static Matrix4 Rotation_Y(float radians);
-	static Matrix4 Rotation_Z(float radians);
-	static Matrix4 Shear(float xy, float xz, float yx, float yz, float zx, float zy);
+	static Matrix4 Rotation_X(double radians);
+	static Matrix4 Rotation_Y(double radians);
+	static Matrix4 Rotation_Z(double radians);
+	static Matrix4 Shear(double xy, double xz, double yx, double yz, double zx, double zy);
 	static Matrix4 ViewTransform(Tuple from, Tuple to, Tuple up);
 
 	// Methods
-	std::vector<float> get_row(int row) const;
-	std::vector<float> get_column(int col) const;
+	std::vector<double> get_row(int row) const;
+	std::vector<double> get_column(int col) const;
 	Tuple get_row_tuple(int row) const;
 	Tuple get_column_tuple(int col) const;
 
 	Matrix3 sub_matrix4(int remove_row, int remove_col) const;
 
-	float minor(int row, int col) const;
-	float cofactor(int row, int col) const;
-	float determinant() const;
+	double minor(int row, int col) const;
+	double cofactor(int row, int col) const;
+	double determinant() const;
 	Matrix4 inverse() const;
 	Matrix4 transpose() const;
